Avoid overflow in rc->this_frame_target * 7
Cast the first operand to int64_t to avoid overflow.
rc->this_frame_target can be as large as INT_MAX.

+++ av1/encoder/rc_utils.h
@@ -144,7 +144,8 @@ static inline int recode_loop_test(AV1_COMP *cpi, int high_limit, int low_limit,
       // Deal with frame undershoot and whether or not we are
       // below the automatically set cq level.
       if (q > oxcf->rc_cfg.cq_level &&
-          rc->projected_frame_size < ((rc->this_frame_target * 7) >> 3)) {
+          rc->projected_frame_size <
+              (((int64_t)rc->this_frame_target * 7) >> 3)) {
         force_recode = 1;
       }
     }
